requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
python-dotenv
azure-identity
azure-keyvault-secrets
